Tags controller with test
