+++ app/controllers/carto/api/visualizations_controller.rb
@@ -5,6 +5,7 @@ require_relative '../../../models/visualization/stats'
 module Carto
   module Api
     class VisualizationsController < ::Api::ApplicationController
+      include VisualizationSearcher
 
       # TODO: compare with older, there seems to be more optional authentication endpoints
       skip_before_filter :api_authorization_required, only: [:index, :vizjson2]
@@ -15,10 +16,6 @@ module Carto
       before_filter :load_visualization, only: [:likes_count, :likes_list, :is_liked, :show, :stats]
       ssl_required :index, :show
       ssl_allowed  :vizjson2, :likes_count, :likes_list, :is_liked
-
-      FILTER_SHARED_YES = 'yes'
-      FILTER_SHARED_NO = 'no'
-      FILTER_SHARED_ONLY = 'only'
 
       def id_and_schema_from_params
         if params.fetch('id', nil) =~ /\./
@@ -58,54 +55,11 @@ module Carto
       end
 
       def index
-        types, total_types = get_types_parameters
         page = (params[:page] || 1).to_i
         per_page = (params[:per_page] || 20).to_i
         order = (params[:order] || 'updated_at').to_sym
-        pattern = params[:q]
-
-        only_liked = params[:only_liked] == 'true'
-        only_shared = params[:only_shared] == 'true'
-        exclude_shared = params[:exclude_shared] == 'true'
-        locked = params[:locked]
-        shared = compose_shared(params[:shared], only_shared, exclude_shared)
-
-        vqb = VisualizationQueryBuilder.new
-            .with_prefetch_user
-            .with_prefetch_table
-            .with_prefetch_permission
-            .with_prefetch_external_source
-            .with_types(types)
-
-        if current_user
-          if only_liked
-            vqb.with_liked_by_user_id(current_user.id)
-          end
-
-          case shared
-          when FILTER_SHARED_YES
-            vqb.with_owned_by_or_shared_with_user_id(current_user.id)
-          when FILTER_SHARED_NO
-            vqb.with_user_id(current_user.id) if !only_liked
-          when FILTER_SHARED_ONLY
-            vqb.with_shared_with_user_id(current_user.id)
-          end
-
-          if locked == 'true'
-            vqb.with_locked(true)
-          elsif locked == 'false'
-            vqb.with_locked(false)
-          end
-        else
-          # TODO: ok, this looks like business logic, refactor
-          subdomain = CartoDB.extract_subdomain(request)
-          vqb.with_user_id(Carto::User.where(username: subdomain).first.id)
-              .with_privacy(Carto::Visualization::PRIVACY_PUBLIC)
-        end
-
-        if pattern.present?
-          vqb.with_partial_match(pattern)
-        end
+        types, total_types = get_types_parameters
+        vqb = query_builder_with_filter_from_hash(params)
 
         # TODO: undesirable table hardcoding, needed for disambiguation. Look for
         # a better approach and/or move it to the query builder
@@ -166,23 +120,6 @@ module Carto
 
       private
 
-      def get_types_parameters
-        # INFO: this fits types and type into types, so only types is used for search.
-        # types defaults to type if empty.
-        # types defaults to derived if type is also empty.
-        # total_types are the types used for total counts.
-        types = params.fetch(:types, "").split(',')
-
-        type = params[:type].present? ? params[:type] : (types.empty? ? nil : types[0])
-        # TODO: add this assumption to a test or remove it (this is coupled to the UI)
-        total_types = [(type == Carto::Visualization::TYPE_REMOTE ? Carto::Visualization::TYPE_CANONICAL : type)].compact
-
-        types = [type].compact if types.empty?
-        types = [Carto::Visualization::TYPE_DERIVED] if types.empty?
-
-        return types, total_types
-      end
-
       def set_vizjson_response_headers_for(visualization)
         # We don't cache non-public vis
         if @visualization.is_publically_accesible?
@@ -199,22 +136,6 @@ module Carto
       def to_hash(visualization)
         # TODO: previous controller uses public_fields_only option which I don't know if is still used
         VisualizationPresenter.new(visualization, current_viewer).to_poro
-      end
-
-      def compose_shared(shared, only_shared, exclude_shared)
-        valid_shared = shared if [FILTER_SHARED_ONLY, FILTER_SHARED_NO, FILTER_SHARED_YES].include?(shared)
-        return valid_shared if valid_shared
-
-        if only_shared
-          FILTER_SHARED_ONLY
-        elsif exclude_shared
-          FILTER_SHARED_NO
-        elsif exclude_shared == false
-          FILTER_SHARED_YES
-        else
-          # INFO: exclude_shared == nil && !only_shared
-          nil
-        end
       end
 
     end
