+++ spec/requests/api/json/visualizations_controller_shared_examples.rb
@@ -36,19 +36,7 @@ shared_examples_for "visualization controllers" do
   end
 
   def factory(user, attributes={})
-    {
-      name:                     attributes.fetch(:name, "visualization #{rand(9999)}"),
-      tags:                     attributes.fetch(:tags, ['foo', 'bar']),
-      map_id:                   attributes.fetch(:map_id, ::Map.create(user_id: user.id).id),
-      description:              attributes.fetch(:description, 'bogus'),
-      type:                     attributes.fetch(:type, 'derived'),
-      privacy:                  attributes.fetch(:privacy, 'public'),
-      source_visualization_id:  attributes.fetch(:source_visualization_id, nil),
-      parent_id:                attributes.fetch(:parent_id, nil),
-      locked:                   attributes.fetch(:locked, false),
-      prev_id:                  attributes.fetch(:prev_id, nil),
-      next_id:                  attributes.fetch(:next_id, nil)
-    }
+    visualization_template(user)
   end
 
   def table_factory(options={})
@@ -407,9 +395,7 @@ shared_examples_for "visualization controllers" do
             password: 'clientex',
         )
 
-        post api_v1_visualizations_create_url(user_domain: @user.username, api_key: @api_key),
-          factory(@user).to_json, @headers
-        vis_1_id = JSON.parse(last_response.body).fetch('id')
+        vis_1_id = create_visualization(@user).id
 
         get api_v1_visualizations_likes_count_url(user_domain: @user.username, id: vis_1_id, api_key: @api_key)
         JSON.parse(last_response.body).fetch('likes').to_i.should eq 0
